feat: use single optimizelyClient if possible
Small performance optimisation that would prevent multiple event dispatches to happen
