+++ lib/flags.ts
@@ -1,108 +1,80 @@
-import optimizely from "@optimizely/optimizely-sdk";
-import { get } from "@vercel/edge-config";
 import { flag } from "flags/next";
-import { dispatchEvent } from "./event-dispatcher";
+import { ensureOptimizely } from "./optimizely";
 import { getShopperFromHeaders } from "./utils";
 
 export const showBuyNowFlag = flag<{
-  enabled: boolean;
-  buttonText?: string;
+	enabled: boolean;
+	buttonText?: string;
 }>({
-  key: "buynow",
-  description: "Flag for showing Buy Now button on PDP",
-  options: [
-    { label: "Hide", value: { enabled: false } },
-    { label: "Show", value: { enabled: true } },
-  ],
-  async decide({ headers }) {
-    const datafile = await get("datafile");
+	key: "buynow",
+	description: "Flag for showing Buy Now button on PDP",
+	options: [
+		{ label: "Hide", value: { enabled: false } },
+		{ label: "Show", value: { enabled: true } },
+	],
+	async decide({ headers }) {
+		let flag = { enabled: false, buttonText: "" };
 
-    if (!datafile) {
-      throw new Error("Failed to retrive datafile from Vercel Edge Config");
-    }
+		try {
+			const client = await ensureOptimizely();
 
-    let flag = { enabled: false, buttonText: "" };
+			if (!client) {
+				throw new Error("Failed to create client");
+			}
 
-    try {
-      const client = optimizely.createInstance({
-        datafile: datafile as object,
-        eventDispatcher: {
-          dispatchEvent,
-        },
-      });
+			const shopper = getShopperFromHeaders(headers);
+			const context = client.createUserContext(shopper);
 
-      if (!client) {
-        throw new Error("Failed to create client");
-      }
+			if (!context) {
+				throw new Error("Failed to create user context");
+			}
 
-      await client.onReady({ timeout: 500 });
-
-      const shopper = getShopperFromHeaders(headers);
-      const context = client.createUserContext(shopper);
-
-      if (!context) {
-        throw new Error("Failed to create user context");
-      }
-
-      const decision = context.decide("buynow");
-      flag = {
-        enabled: decision.enabled,
-        buttonText: decision.variables.buynow_text as string,
-      };
-    } catch (error) {
-      console.error("Optimizely error:", error);
-    } finally {
-      return flag;
-    }
-  },
+			const decision = context.decide("buynow");
+			flag = {
+				enabled: decision.enabled,
+				buttonText: decision.variables.buynow_text as string,
+			};
+		} catch (error) {
+			console.error("Optimizely error:", error);
+		} finally {
+			return flag;
+		}
+	},
 });
 
 export const showPromoBannerFlag = flag<boolean>({
-  key: "showPromoBanner",
-  defaultValue: false,
-  description: "Flag for showing promo banner on homepage",
-  options: [
-    { value: false, label: "Hide" },
-    { value: true, label: "Show" },
-  ],
-  async decide({ headers }) {
-    const datafile = await get("datafile");
+	key: "showPromoBanner",
+	defaultValue: false,
+	description: "Flag for showing promo banner on homepage",
+	options: [
+		{ value: false, label: "Hide" },
+		{ value: true, label: "Show" },
+	],
+	async decide({ headers }) {
+		let flag = false;
 
-    if (!datafile) {
-      throw new Error("Failed to retrieve datafile from Vercel Edge Config");
-    }
+		try {
+			const client = await ensureOptimizely();
 
-    let flag = false;
+			if (!client) {
+				throw new Error("Failed to create client");
+			}
 
-    try {
-      const client = optimizely.createInstance({
-        datafile: datafile as object,
-        eventDispatcher: {
-          dispatchEvent,
-        },
-      });
+			const shopper = getShopperFromHeaders(headers);
+			const context = client.createUserContext(shopper);
 
-      if (!client) {
-        throw new Error("Failed to create client");
-      }
+			if (!context) {
+				throw new Error("Failed to create user context");
+			}
 
-      await client.onReady({ timeout: 500 });
-
-      const shopper = getShopperFromHeaders(headers);
-      const context = client!.createUserContext(shopper);
-
-      if (!context) {
-        throw new Error("Failed to create user context");
-      }
-
-      const decision = context.decide("showpromo");
-      flag = decision.enabled;
-    } catch (error) {
-      console.error("Optimizely error:", error);
-    } finally {
-      return flag;
-    }
-  },
+			const decision = context.decide("showpromo");
+			flag = decision.enabled;
+		} catch (error) {
+			console.error("Optimizely error:", error);
+		} finally {
+			return flag;
+		}
+	},
 });
 
 export const precomputeFlags = [showPromoBannerFlag] as const;
